+++ lib/primitives/takeTwo.test.ts
@@ -1,5 +1,23 @@
 import { assertEquals } from "@std/assert";
 import { any, many, take, takeTwo } from "../main.ts";
+import { parseErrors } from "../../src/errors.ts";
+
+Deno.test("two items", () => {
+  assertEquals(takeTwo.parse("m"), {
+    success: false,
+    message: parseErrors.takeTwoError,
+    position: { line: 1, column: 0 },
+  });
+
+  assertEquals(takeTwo.parse("monad"), {
+    success: true,
+    results: [{
+      value: "mo",
+      remaining: "nad",
+      position: { line: 1, column: 2 },
+    }],
+  });
+});
 
 // Explore a search space
 const oneOrTwoItems = any(take, takeTwo);
